primitives: Make hex optional
Make the `hex` dependency optional. This means not implementing
`Display` for some types if `hex` is not enabled and only implementing
`Debug`.

Note: this is Tobin's commit edited by Martin Habovstiak. The edits are
fixes to depend on `hex?/` rather than `hex/`.

Rationale for this approach: this increases compile time and possibly
even binary size but the other solutions are worse: turning it into a
runtime check (based on serializer being human readable or not) can
cause surprising errors and using
`all(feature = "serde, feature = "hex")` everywhere is tedious and hard
to understand.