DEMO: Make script to/from hex use consensus encoding
I don't know if there is a reason that we do not use consensus encoding
when converting scripts to/from hex strings.

This patch shows the change but does not fix all the call sites that
break - that requires working out the compact var prefix for each
hard coded string we have in the code base.