units: Add sat_per_kvb field to FeeRate
The `FeeRate` has issues with precision. There is a bit of movement in
Core about the same thing. Instead of solving the problem just add an
additional field to the struct that allows us to further play around
with precision. Document as such.

Also there is an off-by-one in the `FeeRate::to_fee` caused by this
imprecision (div by 4 followed by mul by 4).

Fix: #3806