fix(deps): upgrade deps for min version checks
* Standardizing the minimal versions of `serde_json` and `serde_test` in
the workspace. This allows the unstable "direct-minimal-versions" flag
of cargo to be used since it requires one minimal version in a
workspace. Since these are all dev dependencies (including the fuzz
crate) there shouldn't be any change for external consumers.
* Update serde's contraint to the minimum version which the
workspace has been using due to implicit upgrades. Attempts were first
made to see if 1.0.103 could be used, but at the very least, would
require a fair amount of lint settings to be overridden due to the old
code.
* Arbitrary 1.4.0 does not work with the current 1.63.0 MSRV due to some
core::error usage.