chore(deps): bump dev serde versions
Standardizing the minimal versions of `serde_json` and `serde_test` in
the workspace. This allows the unstable "direct-minimal-versions" flag
of cargo to be used since it requires one minimal version in a
workspace. Since these are all dev dependencies (including the fuzz
crate) there shouldn't be any change for external consumers.