build: use direct-minimal-versions for lockfile
Instead of relying on "cargo check" to conservatively upgrade the
minimal versions lockfile, resolve it on any change to ensure the
direct minimal dependencies of the workspace are actually tested.

The issue with "cargo check" is that any version downgrades of direct
dependencies are not automatically reflected in the lockfile. They would
have to be hand set by developers. The unstable direct-minimal-versions
flag of cargo does this for us. It also limits minimal versions to just
direct dependencies. This allows us to check our minimal versions
without requiring our dependencies to also check theirs.