build: use direct-minimal-versions for lockfile
Adding the unstable direct-minimal-versions flag insteaf of relying on
just "cargo check" to conservatively upgrade the minimal versions
lockfile.

The issue with "cargo check" by itselft is that any version downgrades
of direct dependencies are not automatically reflected in the lockfile.
They would have to be hand set by developers. The direct-minimal-versions
flag of cargo does this for us, overriding just that aspect of the
conservative resolver of check.