Add ops::Sub to BlockTime
Implement Sub for BlockTime

The tricky part here is deciding on the output type, since it could be
Duration, u32, or i64.

In this implementation, we chose to return i64
to allow both negative and positive differences.

Also, since BlockTime uses u32, any difference between two timestamps
will fit in an i64.