Delete `FormatOptions`
`FormatOptions` was a workaround to support arbitrary `fmt::Write` in
legacy API. That API is now removed so `FormatOptions` is no longer
needed because we can just accept `Formatter` and pull the options from
itself.

This commit removes the struct and the argument from `fmt_satoshi_in`,
changing `f` from `fmt::Write` to `fmt:: Formatter`.