Move number normalization from Amount to Decimal
Before a decimal number is displayed it is normalized to not end with 0
unless it's 0. This was performed in amount formatting but it applies to
all numbers. The change was intentionally split up to make review
easier.

This moves the computation from a single branch and makes it
unconditional. This is valid because other branches keep
`num_after_decimal_point` as 0 and the condition in the moved code does
0 check, returning 0 which is the valid value for the other branches.