Move rounding from `Amount` to `Decimal`
When displaying a decimal number rounding may be required to satisfy
`precision` field on formatter. This implies rounding is also a general
concept, not tied to `Amount` and needs to be moved to decimal as well.

This commit moves rounding but because decimal accepts numbers before
and after decimal point it needs to be modified a bit to account for the
possibility that the number before decimal point increases because of
rounding. E.g. 0.999 Rounded to two decimal places is 1.00.