Support tweaking of `FeeRate` formatting
So far `FeeRate` had very sketchy implementation that relied on
`alternate` to select the unit, which was surprising and limiting.
Also the sat/vB implementation computed ceiling but also added `.00`
which looked like the value was exact. Further, the unit was always
shown in case of sat/vB and never shown in case of sat/kwu.

To fix it, this commit adds a new `Display` type similar to the one in
`amount` which can be used to tweak how it's displayed. Because
`FeeRate` so far supports multiple rounding modes the support of these
modes was added to the `Decimal` type and the `Decimal` type was used to
implement `Display`. The unit is displayed by default to avoid confusion
and can be hidden with a method call.

This doesn't modify the `Display` impl yet because it's more complicated
and easier to review separately.