Change `Display` of `FeeRate` to show unit
As mentioned in previous commit, the `Display` impl was inconsistent and
problematic in multiple ways. The previous commit introduced an
alternative API that can tweak the formatting but didn't touch the
`Display` impl. This one does but to do that `FromStr` had to be changed
as well.

This changes the `Display` impl to always show the unit and also
delegates existing `sat/vB` code to the newly-added API, so that decimal
formatting is more sensible. This change is very obvious in output and
breaks parsing which is intended. So to ensure our roundtrip tests keep
working this also modifies `FromStr` to require the unit similarly to
how `Amount` requires denomination.

For simplicity, the code currently doesn't support sat/vB parsing but
does detect the attempt and reports the error properly. This can be
added later.