Make `Decimal` more internally consistent
When a number has no decimals, setting the value and rounding doesn't
make sense since they will not be used. Conversely, if the number after
decimal point is 0 then setting number of decimals makes no sense since
it would be normalized to 0. So far the code dealing with these cases
was setting useless values to those fields.

This change refactors the code to use a single filed which is an
`Option` to represent the number after decimal point being present or
not. So either all values need to be provided or none of them. To
improve clarity, `NonZero` integers are used to enforce that the values
are not zero.

Since the code internally transforms the value another, similar type is
used. But this time it's an enum to represent the possibility that
there's no actual number and it's just trailing zeros requested by
formatter.

Using `NonZero` also uncovered an API edge case: setting inconsistent
values to those fields can lead to broken behavior. I don't see a
straightforward solution, and the API is internal, so keeping it as
panic seems reasonable.

During the refactor I also noticed overflow when consumer entered
absurdly large precision was not handled. Since it only affects padding
and the resulting number would be larger than any possible `width`
provided by the formatter the correct fix is to use `saturating_add`.
That way the ("incorrectly") computed number width will be larger or
equal to `width` leading to zero padding and if it was possible to
compute larger number width it'd still be larger than `width` leading to
zero padding.