primitives: Clarify usage of is_satisfied_by_
For relative time locks we provide an API for checking if a relative
lock time is satisfied height and one for time however we do not clearly
state what satisfied means or exactly what the function arguments means.

For `relative::LockTime::is_satisfied_by_height`:

Explicitly make the `height` argument be the delta between the chain tip
and height that the utxo was mined. This means that because the next
block will increase the block height we add 1 when doing the comparison.

For `relative::LockTime::is_satisfied_by_time`:

Document that the `time` argument is the delta between the MTP of the
chain tip and the mining time of the UTXO. Note 'mining time' is defined
in BIP-68 as the MTP of the previous block. When Core checks the next
block the MTP of the previous block is the MTP of the chain tip hence no
manipulation is needed by us during comparison (in contrast to the +1 in
the height function).