bip32: return error when attempting to derive past maximum depth
The corresponding PR on master is #4387, but this doesn't really
resemble that PR. Rather than changing all the error enums, this just
adds a new variant to the #[non_exhaustive] bip32::Error enum and
returns that when adding 1 to 255 when deriving child keys.

This is therefore not an API break and can be released in a minor
version. Although it does change the error return on private derivation
from a "succeeds except with negligible probability" to "there is an
error path you may need to check for".

Fixes #4308