Rename absolute locktime value types
The lock time values for absolute locks are absolute height and a UNIX
timestamp. The encoding on chain uses a threshold because all sane
values of lock-by-time and lock-by-height values fall nicely on either
side of this threshold.

We also have a `BlockTime` and `BlockHeight` type in `primitives` that is
identical except does not enforce the threshold.

Within the locktime modules the purpose of the height and time value is
to add type safety and abstract over the threshold encoding thing.

Outside of the locktime module the types are to help devs know what they
are and provide type safety.

The compiler handles type safety and since the threshold abstraction is
invisible for all sane values we can re-use the same names as the two
types that do not enforce the threshold. This seems racy at first blush
but since the lock time values and the other types are exactly the same
concept and the compiler catches misuse this change _removes_ developer
confusion because devs can forget about why there are two types and just
follow the compiler if conversion is needed.