Fix Psbt preimage keys in serde test
The preimage values in the serde Psbt don't actually correspond to the
hash keys they should in the serde test. This doesn't cause an error
currently because the derived serde implementation doesn't enforce their
sanity during deserializaqtion, but it will when the serde
implementation is modified to use the BIP-174 format.