refactor!: move `p2p::Magic` to `network`
This is an implicit step in moving `p2p` to its own crate. Ideally
nothing in `bitcoin` depends on `p2p`, but `p2p::Magic` is used as a
constructor for `Network`. In my opinion, network "magic" and the
cryptocurrency network are somewhat interchangable, in that a network
magic is meaningless without an underlying network. This PR relocates
the `Magic` struct into `network` along with the associated test.