internals: add macro for defining an error
Errors with no sources have the same implementation body. This can be
automated for many error types across the crate.

For the sake of easy review, I replaced the direct occurances where an
error explicitly defined the `source` method. There are also empty
`std::error::Error` implementations that could make use of the macro,
but I omitted them to preserve the exact API that was present before.