refactor!: move `p2p` into a crate
New peer-to-peer messages are added to Bitcoin Core occasionally.
Including `p2p` in the general `bitcoin` crate restricts changes in its
API to the SemVer semantics of `bitcoin`. This would generally prevent
new peer-to-peer messages in the `p2p` module. Except for a few types,
`bitcoin` does not depend on `p2p`, so `p2p` can have an independent
release cycle of `bitcoin`.

Caveats:
- The `Network` type included some methods that depended on
  `p2p::Magic`. These are removed, as `TryFrom` is implemented to and
  from magic/network
- `Encodable` is no longer defined within `p2p`, so encodable vector
  types violate the orphan rule when attempting to implement `Encodable`
  for `Vec<T: Encodable>`. Instead of introducing a generic
  implementation for `Vec<T>`, I used wrapper types and implemented
  encodable for those
- `Network` is no longer defined at the crate level so the match arm
  becomes non-exhaustive. This requires `TryFrom` for Network to Magic
  now
- The only types available in `no-std` were `Magic` and `ServiceFlags`.
  It simplifies the crate greatly to assume a standard environment. One
  can implement their own concept of magic without this crate.
- `ServiceFlags` appeared to have a duplicate test