bitcoin: secp256k1 global-context feature re-export
Often we want to have the global-context feature in secp256k1 without
having to add manually the secp256k1 dependency and enabling the
global-context feature.

Having the ability to do that directly from bitcoin without having to
add secp256k1 and do the whole tango of tightly coupling the two
dependecy versions together, e.g. bitcoin 0.32.x and secp256k1 0.29.x
would be really nice and would also simplify a lot code maintainability
for anyone who depends on bitcoin.

This needs to be backported to 0.32.x, which I'll gladly do as well.