Change InputWeightPrediction types to u32 from usize
It's possible to cause an overflow when calling the
`InputWeightPrediction` constructors with extreme values.  Furthermore,
even if the constructors do not overflow, it's then possible for
associated methods such as `total_weight()` and `witness_weight()` to
overflow.

To solve this without grossly changing the API, use `u32` and then
within associated methods perform arithmetic using `u64`.  It's still
possible to overflow in the constructors however since they return `u32`
values, so the arithmetic is changed to saturating in the next commit.