Add FromStr to FeeRate
Allow parsing of strings to FeeRate using either sat/kwu or sat/vB
notation.  Also make a special case for if zero is used which needs no
notation since zero is zero in every basis.  Also provide a catch if
no space was provided between the digits and rate notation.