Change param datatype for InputWeightPrediction functions
Any value for `bytes_to_grind` that is larger than 62 will cause the
subtraction for `der_signature_size` to be negative.  However,
`der_signature_size` must be positive when passed to the
`InputWeightPrediction` constructor.  Therefore there is no value in
allowing values of size `usize` to be passed due to any value greater
than 62 is invalid.

It would be preferable to use `.into()` instead of `as` when casting to
`usize` for the constructor call, however these functions are const
context and therefore `.into()` is not readily available.